Refactor to livesketch
