+++ src/livesketch/local_simhash.py
